requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pdfplumber>=0.10.2
FPDF2>=2.7.4
Pillow>=10.0.0
python-dateutil>=2.8.2
flask>=2.3.3
flask-cors>=4.0.0
